Update room screen
fix reservation and reservation booking
